Move Annotation class def and update docstrings
